integration: extract utils into module
